+++ src/services/actions/ingredients.js
@@ -1,0 +1,28 @@
+import Api from "../../api/api";
+
+export const GET_INGREDIENT_REQUEST = "GET_INGREDIENT_REQUEST";
+export const GET_INGREDIENT_SUCCESS = "GET_INGREDIENT_SUCCESS";
+export const GET_INGREDIENT_FAILED = "GET_INGREDIENT_FAILED";
+
+const api = new Api();
+
+export function getIngredients() {
+  return function (dispatch) {
+    dispatch({
+      type: GET_INGREDIENT_REQUEST,
+    });
+    api
+      .getIngredientsList()
+      .then(({ data }) => {
+        dispatch({
+          type: GET_INGREDIENT_SUCCESS,
+          data,
+        });
+      })
+      .catch((e) => {
+        dispatch({
+          type: GET_INGREDIENT_FAILED,
+        });
+      });
+  };
+}
