--- src/utils/constructorContext.js
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const ConstructorContext = createContext([]);
